Use prometheus labels properly
Previously I was creating one metric for every gmail label. This changes
it so we're creating just two metrics (total and unread), with gmail
lables provided as prometheus labels.

It's not completely perfect because you can't reliably SUM gmail labels
because one email might have lots of labels (so would be counted
multiple times). But it still feels more correct than having multiple
metrics.

As a nice side-benefit, this makes the code massively simpler and easier
to follow :)